.gitignore - ignore output file
10Aug11.csv - Larger sample size (note: this looks wrong in GitHub app,
double check on website)

project.clj - add dependency on clojure-contrib for the math library

core.clj -
  *  Change col-heads from a map to a vector. Now instead of looking up
where a fields is we just zipmap the field vector with the row vector.
  *  Change cleanse-csv to take a key-value vector instead of two
arguments. This makes preserving the hash-map structure of the record
table easier.
  *  csv-tuples is now csv-records. The records table is now a sequence
of hash-maps. Since order within a row is no longer important, the
field list for filtering is now a set. Other changes were made to
create the sequence-of-hash-maps structure.
  *  normalize-datetime automatically works on datetime.
  *  slope is now velocity and it destructs a record using the datetime
and pressure keys.
  *  seq-slope is gone
  *  motion uses pressure and velocity to determine if the bird is
ascending, descending, on the surface, or at the bottom. Needs better
heuristics for more accurate results.
  *  strr is recursive str. Converts all scalars inside of nested
collections to string for writing to csv.
  *  analyze-dives expands the record table from datetime and pressure
to those plus velocity and motion. Still need to add dive #.
  *  write-csv rejiggers the record table and spits it to a file using
csv/write-csv.